stucture the project into files contains fuctions
