+++ src/js/clearAllElement.js
@@ -1,0 +1,8 @@
+const clearAllElement = () => {
+  const scoresContainer = document.querySelector('ul');
+  while (scoresContainer.firstChild) {
+    scoresContainer.firstChild.remove();
+  }
+};
+
+export default clearAllElement;
